+++ app/services/CollectionService.scala
@@ -1,7 +1,10 @@
 package services
 
+import java.util.Date
+
 import api.Permission.Permission
-import models.{User, UUID, Dataset, Collection}
+import models.{Collection, Dataset, UUID, User}
+
 import scala.collection.mutable.ListBuffer
 import scala.util.Try
 
@@ -71,6 +74,11 @@ trait CollectionService {
   def listUser(limit: Integer, user: Option[User], showAll: Boolean, owner: User): List[Collection]
 
   /**
+    * Return a list of collections the user has created in the trash.
+    */
+  def listUserTrash(limit: Integer, user: Option[User], showAll: Boolean, owner: User): List[Collection]
+
+  /**
    * Return a list of collections the user has created with matching title.
    */
   def listUser(limit: Integer, title: String, user: Option[User], showAll: Boolean, owner: User, exact: Boolean): List[Collection]
@@ -79,6 +87,17 @@ trait CollectionService {
    * Return a list of collections the user has created starting at a specific date.
    */
   def listUser(date: String, nextPage: Boolean, limit: Integer, user: Option[User], showAll: Boolean, owner: User): List[Collection]
+
+  /**
+    * Return a list of collections the user has created starting at a specific date in the trash.
+    */
+  def listUserTrash(date: String, nextPage: Boolean, limit: Integer, user: Option[User], showAll: Boolean, owner: User): List[Collection]
+
+
+  /**
+    * Return a number of collections the user has created in trash.
+    */
+  def listUserTrash(user : Option[User],limit : Integer ) : List[Collection]
 
   /**
    * Return a list of collections the user has access to starting at a specific date with matching title.
@@ -133,6 +152,10 @@ trait CollectionService {
    * Update description of the dataset
    */
   def updateDescription(id: UUID, description: String)
+
+  def addToTrash(id : UUID, dateMovedToTrash : Option[Date])
+
+  def restoreFromTrash(id : UUID, dateMovedToTrash : Option[Date])
 
   /**
    * Delete collection and any reference of it